Fril re exhibit

+++ app/src/Controller/UsersController.php
@@ -3,6 +3,7 @@ namespace App\Controller;
 
 use App\Controller\AppController;
 use Cake\Event\Event;
+use Cake\ORM\TableRegistry;
 
 /**
  * Users Controller
@@ -11,6 +12,7 @@ use Cake\Event\Event;
  */
 class UsersController extends AppController
 {
+    public $user_editable_key = ['username', 'password'];
 
     /**
      * Index method
@@ -19,12 +21,17 @@ class UsersController extends AppController
      */
     public function index()
     {
+        if (!$this->is_admin) {
+            return $this->redirect(['action'=>'edit']);
+        }
         $this->paginate = [
             'contain' => ['Groups', 'Licenses']
         ];
         $users = $this->paginate($this->Users);
 
         $this->set(compact('users'));
+        $this->set('current_user', $this->current_user);
+        $this->set('current_user_shop_account', $this->current_user_shop_account);
         $this->set('_serialize', ['users']);
     }
 
@@ -37,11 +44,19 @@ class UsersController extends AppController
      */
     public function view($id = null)
     {
-        $user = $this->Users->get($id, [
-            'contain' => ['Groups', 'Licenses', 'UserShopAccount', 'UserShopApplyOrderTmpl', 'UserShopEvaluateTmpl', 'UserShopReceiveFeeTmpl', 'UserShopRelistCron', 'UserShopShipTmpl']
-        ]);
+        if (!empty($id)) {
+            if (!$this->is_admin) {
+                return $this->redirect(['action'=>'edit']);
+            }
+            $user = $this->Users->get($id, [
+                'contain' => ['UserShopAccounts','Groups']
+            ]);
+        } else {
+            $user = $this->current_user;
+        }
 
-        $this->set('user', $user);
+        $this->set('current_user', $user);
+        $this->set('current_user_shop_account', $this->current_user_shop_account);
         $this->set('_serialize', ['user']);
     }
 
@@ -52,8 +67,11 @@ class UsersController extends AppController
      */
     public function add()
     {
+        if (!$this->is_admin) {
+            return $this->redirect(['action'=>'edit']);
+        }
         $user = $this->Users->newEntity();
-        if ($this->request->is('post')) {
+        if ($this->request->is(['post', 'put'])) {
             $user = $this->Users->patchEntity($user, $this->request->data);
             if ($this->Users->save($user)) {
                 $this->Flash->success(__('The user has been saved.'));
@@ -66,6 +84,8 @@ class UsersController extends AppController
         $groups = $this->Users->Groups->find('list', ['limit' => 200]);
         $licenses = $this->Users->Licenses->find('list', ['limit' => 200]);
         $this->set(compact('user', 'groups', 'licenses'));
+        $this->set('current_user', $this->current_user);
+        $this->set('current_user_shop_account', $this->current_user_shop_account);
         $this->set('_serialize', ['user']);
     }
 
@@ -78,22 +98,43 @@ class UsersController extends AppController
      */
     public function edit($id = null)
     {
-        $user = $this->Users->get($id, [
-            'contain' => []
-        ]);
+        if (!empty($id)) {
+            if (!$this->is_admin) {
+                return $this->redirect(['action'=>'edit']);
+            }
+            $user = $this->Users->get($id);
+        } else {
+            $user = $this->current_user;
+        }
+
         if ($this->request->is(['patch', 'post', 'put'])) {
+            if (!$this->is_admin) {
+                foreach ($this->request->data as $key => $val) {
+                    if (!in_array($key, $this->user_editable_key)) {
+                        unset($this->request->data[$key]);
+                    }
+                }
+            }
             $user = $this->Users->patchEntity($user, $this->request->data);
             if ($this->Users->save($user)) {
-                $this->Flash->success(__('The user has been saved.'));
+                $this->Flash->success('設定を保存しました。');
 
-                return $this->redirect(['action' => 'index']);
+                if ($this->is_admin) {
+                    return $this->redirect(['action' => 'index']);
+                } else {
+                    return $this->redirect(['action' => 'edit']);
+                }
             } else {
-                $this->Flash->error(__('The user could not be saved. Please, try again.'));
+                $this->Flash->error('設定の保存に失敗しました。');
             }
         }
+
         $groups = $this->Users->Groups->find('list', ['limit' => 200]);
         $licenses = $this->Users->Licenses->find('list', ['limit' => 200]);
         $this->set(compact('user', 'groups', 'licenses'));
+        $this->set('is_admin', $this->is_admin);
+        $this->set('current_user', $this->current_user);
+        $this->set('current_user_shop_account', $this->current_user_shop_account);
         $this->set('_serialize', ['user']);
     }
 
@@ -106,6 +147,9 @@ class UsersController extends AppController
      */
     public function delete($id = null)
     {
+        if (!$this->is_admin) {
+            return $this->redirect(['action'=>'edit']);
+        }
         $this->request->allowMethod(['post', 'delete']);
         $user = $this->Users->get($id);
         if ($this->Users->delete($user)) {
@@ -130,9 +174,17 @@ class UsersController extends AppController
 
             if ($user) {
                 $this->Auth->setUser($user);
+                $user_shop_account_id = $this->Cookie->read('usai');
+                $user_shop_accounts_tbl = TableRegistry::get('user_shop_accounts');
+                if (empty($user_shop_account_id)) {
+                    $this->current_user_shop_account = $user_shop_accounts_tbl->find()->where(['user_id'=> $user['id']])->first();
+                    $this->Cookie->write('usai', $this->current_user_shop_account->id);
+                } else {
+                    $this->current_user_shop_account = $user_shop_accounts_tbl->get($user_shop_account_id);
+                }
                 return $this->redirect($this->Auth->redirectUrl());
             }
-            $this->Flash->error(__('Invalid username or password, try again'));
+            $this->Flash->error('ログイン情報が間違っています');
         }
     }
